Corrects various small errors
In the route folder I was unable to launch the server without it crashing due to certain elements not being defined. This was updated with new route lines being added.

In the DAO file the database variable is 'database' not 'internships', so I updated those variables. 'ObjectId' still remains undefined however.